# Use a flat layout for TokenBuffer